+++ packages/node/src/fork-dev-server.ts
@@ -15,6 +15,7 @@ export function forkDevServer(options: {
   require_: NodeRequire;
   entrypoint: string;
   meta: Meta;
+  printLogs?: boolean;
 
   /**
    * A path to the dev-server path. This is used in tests.
@@ -59,8 +60,18 @@ export function forkDevServer(options: {
         : undefined,
       NODE_OPTIONS: nodeOptions,
     }),
+    stdio: options.printLogs ? 'pipe' : undefined,
   };
   const child = fork(devServerPath, [], forkOptions);
+
+  if (options.printLogs) {
+    child.stdout?.on('data', data => {
+      console.log(`stdout: ${data}`);
+    });
+    child.stderr?.on('data', data => {
+      console.error(`stderr: ${data}`);
+    });
+  }
 
   checkForPid(devServerPath, child);